requirements.txt
Click==7.0
colorama==0.4.1
python-git==2018.2.1
PyYAML==3.13
Send2Trash==1.5.0
